+++ package.json
@@ -13,9 +13,11 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.395.0",
     "axios": "^1.4.0",
+    "dompurify": "^3.0.5",
     "eslint": "8.47.0",
     "eslint-config-next": "13.4.13",
     "http-errors": "^2.0.0",
+    "jsdom": "^22.1.0",
     "mongoose": "^7.4.3",
     "next": "13.4.13",
     "next-auth": "^4.23.1",
